added dialog box

+++ src/app/component/list-employee/list-employee.component.ts
@@ -1,7 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, signal } from '@angular/core';
 import { EmployeeService } from '../../service/employee.service';
 import { Employee } from '../../model/employee';
 import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent } from '../delete-dialog/delete-dialog.component';
 
 @Component({
   selector: 'app-list-employee',
@@ -15,7 +17,8 @@ export class ListEmployeeComponent implements OnInit {
   displayedColumns: string[] = ['id', 'firstName', 'lastName', 'emailId'];
 
 
-  constructor(private employeeSvc: EmployeeService, private router: Router) { }
+
+  constructor(private dialog: MatDialog, private employeeSvc: EmployeeService, private router: Router) { }
 
   ngOnInit(): void {
     this.getEmployees();
@@ -32,9 +35,31 @@ export class ListEmployeeComponent implements OnInit {
   }
 
   deleteEmployee(arg0: any) {
-    this.employeeSvc.deleteById(arg0).subscribe((data: any) => {
-      this.getEmployees();
+
+    // logic for Dialog Popup here
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      data: {
+        messageTxt: "Sure to override record?",
+        cancelTxt: "No",
+        confirmTxt: "Yes"
+      }
     });
+
+    dialogRef.afterClosed().subscribe((result: any) => {
+      console.log(`Dialog result: ${result}`);
+      if (result) {
+        this.employeeSvc.deleteById(arg0).subscribe((data: any) => {
+          this.getEmployees();
+        });
+      }
+      else {
+        // do nothing
+        console.log("User clicked No");
+      }
+    })
+
+
+
   }
 
   displayEmployee(arg0: any) {
